side bar and reactivity

+++ Models/Bed.cs
@@ -10,8 +10,8 @@ namespace GardenBoxer.Models
     public string Description { get; set; }
     public double Width { get; set; }
     public double Height { get; set; }
-    public int BedX { get; set; }
-    public int BedY { get; set; }
+    public double BedX { get; set; }
+    public double BedY { get; set; }
     public DateTime DatePlanted { get; set; }
     public DateTime DateFertilized { get; set; }
     public string Img { get; set; }
